requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
robotframework==7.2
urllib3==2.3.0